Split the Asset File Serving to Routing\AssetFileDispatcher
